Apply microphone config into preview component.
