+++ osu.Game.Rulesets.Karaoke/Screens/Config/Sections/Input/MicrophoneSettings.cs
@@ -4,12 +4,16 @@
 using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Game.Rulesets.Karaoke.Configuration;
+using osu.Game.Rulesets.Karaoke.Screens.Config.Previews;
+using osu.Game.Rulesets.Karaoke.Screens.Config.Previews.Input;
 
 namespace osu.Game.Rulesets.Karaoke.Screens.Config.Sections.Input
 {
     public class MicrophoneSettings : KaraokeSettingsSubsection
     {
         protected override string Header => "Microphone";
+
+        public override SettingsSubsectionPreview CreatePreview() => new MicrophoneDevicePreview();
 
         [BackgroundDependencyLoader]
         private void load()
